intiialized front end sockets, got it successfully communicating with server
